+++ docs/changelog.md
@@ -15,6 +15,10 @@ Types of changes
 
 ## Unreleased
 
+Added:
+
+- Add UI option to **re-match existing segments to activities**. This is useful when there is an update to the segment matching algorithm and one wants to try that. (GH-413)
+
 Fixed:
 
 - Fix **segment matching** for cases where the activity is sparse. This might happen when the GPS tracker uses only a few samples or the speed is very high. (GH-413)